Remove `include/grpcpp/impl/codegen/client_unary_call.h` (#31058)
* Remove `include/grpcpp/impl/codegen/client_unary_call.h`

* Automated change: Fix sanity tests

* Fix
